backend/requirements.txt
fastapi==0.116.1
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.13.5
pandas==2.2.2
numpy<2.0
google-cloud-bigquery==3.36.0
google-auth==2.40.3
google-cloud-aiplatform>=1.71.1
python-multipart==0.0.9
alembic==1.13.1
pinecone==7.3.0
sec-api==1.0.32
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.40.1
scikit-learn==1.4.2
scipy==1.13.0
pydantic-settings==2.3.4
pgvector==0.2.0
